moved classes from sdk to __init__. closes #47 #48 #49 and #50.
